requirements.txt
beautifulsoup4==4.13.4
builtwith==1.3.4
dnspython==2.8.0
geoip2==5.1.0
pyfiglet==1.0.4
python-dotenv==1.1.1
python-whois==0.9.5
requests==2.26.0
shodan==1.31.0
termcolor==3.1.0
tqdm==4.67.1
waybackpy==3.0.6
zoomeye==3.0.0
importlib_resources>=5.0.0
PyPDF2==3.0.1
python-docx==1.1.0
Pillow>=11.0.0